Refactor a few things, repair a few edge cases
This PR is a bit of a mixed bag. It contains the changes I had to make to make it work for us. These are:

Deal with `nil` interfaces and slices out-of-bounds. Nothing too bad, just two simple checks added.

Check the _last_ parameter of a function for whether it's a callback. This facilitates signatures where the first parameter is a context, such as `DoThing(context.Context, func(xyz any) error)`.

Check for "method called on receiver even though a parameter with the same type exists". This requires a bit
of an explanation. We have an abstraction for database access that has a way to run a callback function in a
transaction. The callback takes a instance of the abstraction as its paramter, so the signature is roughly:

```go
type DB struct{}

func (*DB) RunTx(context.Context, func(*DB) error) error
```

We want to catch accidental calls to the outer `DB` where the inner was meant, so this should cause an
error:

```go
db := DB{}

err := db.RunTx(context.TODO(), func(d *DB) error {
	return db.DoSomething() // Error: not using parameter `d` but outer `db`
}
```

An example of a similar existing pattern is `*testing.T`'s `Run` method: it receives as its parameters a
name for the subtest and a function that itself takes a `*testing.T` as its parameter. There, we want to
prevent accidental calls to the outer `t`'s methods.

Next, I changed the code that recognizes "This is a captured value" so that it respects parameter names and
does not trip on code like this:

```go
err := db.RunTx(context.TODO(), func(db *DB) error {
	return db.DoSomething() // Should _not_ error because `db` parameter shadows the capture
}
```

The last change is for the resolution of interfaces. We sometimes have situations like this:

```go
type SomeIfaceA interface {
	MethodA()
}

type SomeIfaceB interface {
	SomeIfaceA
	MethodB()
}

DoThing(func(x SomeIfaceA) {
	y, ok := x.(SomeIfaceB)
	if !ok {
		return
	}

	y.MethodB() // Should not error because `MethodB` is not reachable through `x`.
})
```

Previously, the code errored on the call to `y.MethodB()` with the message that `y` implements `SomeIfaceA`. That is true, but in this situation not helpful because `MethodB` is not available on the value of type `SomeIfaceA`. We now check the fully qualified name of the function being called in a call expression against the method sets of all callback parameters. If the function is not part of the method set of any parameter, we don't report an error.